Make external vector type non-opaque (#4)
This allows matching against the type in guards, without dialyzer spitting "Call does not have expected opaque term ..." errors.

This goes in line with [this change on `MapSet`](https://github.com/elixir-lang/elixir/pull/11917/files). For more context on the error, see [this forum post](https://elixirforum.com/t/dialyzer-complaint-about-mapset-member-not-getting-proper-type-as-argument-possible-specs-bug-in-mapset/20780/17).